Cluster/requirements.txt
numpy
scikit-learn
matplotlib
seaborn
torch
torchvision
pillow
tqdm
